Restructure test files and separate application under test.
