+++ src/test/java/com/societegenerale/aut/main/IInterfaceWithIncorrectName.java
@@ -1,0 +1,4 @@
+package com.societegenerale.aut.main;
+
+public interface IInterfaceWithIncorrectName {
+}
