+++ src/test/java/com/societegenerale/aut/main/InterfaceWithCorrectName.java
@@ -1,0 +1,4 @@
+package com.societegenerale.aut.main;
+
+public interface InterfaceWithCorrectName {
+}
