+++ src/test/java/com/societegenerale/aut/main/ObjectWithJavaUtilGregorianCalendar.java
@@ -1,4 +1,4 @@
-package com.societegenerale.commons.plugin.rules.classesForTests;
+package com.societegenerale.aut.main;
 
 import java.util.GregorianCalendar;
