+++ src/test/java/com/societegenerale/aut/main/ObjectWithStandardStream.java
@@ -1,0 +1,8 @@
+package com.societegenerale.aut.main;
+
+public class ObjectWithStandardStream {
+
+    public void doGreatStuff() {
+        System.err.println("I don't know..");
+    }
+}
