+++ src/test/java/com/societegenerale/aut/main/TotallyGoodInterfaceName.java
@@ -1,0 +1,4 @@
+package com.societegenerale.aut.main;
+
+public interface TotallyGoodInterfaceName {
+}
