+++ src/test/java/com/societegenerale/aut/test/TestClassWithAutowiredField.java
@@ -1,4 +1,4 @@
-package com.societegenerale.commons.plugin.rules.classesForTests;
+package com.societegenerale.aut.test;
 
 import org.springframework.beans.factory.annotation.Autowired;
