+++ src/test/java/com/societegenerale/aut/test/TestClassWithIgnoreAtMethodLevel.java
@@ -1,4 +1,4 @@
-package com.societegenerale.commons.plugin.rules.classesForTests;
+package com.societegenerale.aut.test;
 
 import org.junit.Ignore;
 import org.junit.Test;
@@ -6,11 +6,12 @@ import org.junit.Test;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.fail;
 
-@Ignore
-public class TestClassWithIgnoreAtClassLevel {
+
+public class TestClassWithIgnoreAtMethodLevel {
 
     @Test
-    public void someTestUsingAssertJ() {
+    @Ignore
+    public void someIgnoredTestWithoutAComment() {
 
         assertThat(true).isTrue();
