+++ src/test/java/com/societegenerale/aut/test/TestClassWithInjectedField.java
@@ -1,4 +1,4 @@
-package com.societegenerale.commons.plugin.rules.classesForTests;
+package com.societegenerale.aut.test;
 
 import javax.inject.Inject;
