+++ src/test/java/com/societegenerale/aut/test/TestClassWithJunit5Asserts.java
@@ -1,4 +1,4 @@
-package com.societegenerale.commons.plugin.rules.classesForTests;
+package com.societegenerale.aut.test;
 
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
