+++ src/test/java/com/societegenerale/aut/test/TestClassWithOutJunitAsserts.java
@@ -1,15 +1,13 @@
-package com.societegenerale.commons.plugin.rules.classesForTests;
+package com.societegenerale.aut.test;
 
-import org.junit.Ignore;
 import org.junit.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.fail;
 
-public class TestClassWithIgnoreAtMethodLevelWithComment {
+public class TestClassWithOutJunitAsserts {
 
     @Test
-    @Ignore("here's a comment explaining why it's ignored")
     public void someTestUsingAssertJ() {
 
         assertThat(true).isTrue();
