+++ src/test/java/com/societegenerale/aut/test/TestClassWithStandardStream.java
@@ -1,4 +1,4 @@
-package com.societegenerale.commons.plugin.rules.classesForTests;
+package com.societegenerale.aut.test;
 
 public class TestClassWithStandardStream {
