+++ src/test/java/com/societegenerale/aut/test/specificCase/DummyClassToValidate.java
@@ -1,0 +1,4 @@
+package com.societegenerale.aut.test.specificCase;
+
+public class DummyClassToValidate {
+}
