--- src/test/java/com/societegenerale/commons/plugin/rules/classesForTests/InterfaceWithCorrectName.java
@@ -1,4 +1,0 @@
-package com.societegenerale.commons.plugin.rules.classesForTests;
-
-public interface InterfaceWithCorrectName {
-}
