--- src/test/java/com/societegenerale/commons/plugin/rules/classesForTests/TotallyGoodInterfaceName.java
@@ -1,4 +1,0 @@
-package com.societegenerale.commons.plugin.rules.classesForTests;
-
-public interface TotallyGoodInterfaceName {
-}
