--- src/test/java/com/societegenerale/commons/plugin/rules/classesForTests/specificCase/DummyClassToValidate.java
@@ -1,4 +1,0 @@
-package com.societegenerale.commons.plugin.rules.classesForTests.specificCase;
-
-public class DummyClassToValidate {
-}
